1）修复了KeepAlive的问题  2）移除了third_party目录下的文件，采用gitmodules进行管理
- 修复时间变量类型不匹配问题：将uint64_t改为uint32_t以匹配ports_get_epoch_time()返回类型
- 添加无符号整数下溢保护，防止keepalive超时计算错误
- 将keepalive超时从30秒缩短到10秒，更快检测连接问题
- 增强调试日志，添加时间戳便于问题排查
- 移除third_party目录中的所有文件，改用git submodules管理第三方依赖
- 添加libsrtp、mbedtls、usrsctp作为子模块，便于版本管理和更新

+++ include/peer_connection.hpp
@@ -170,8 +170,8 @@ private:
     uint32_t remote_vssrc_;
     
     // Keepalive management
-    uint64_t next_keepalive_time_;
-    uint64_t time_of_last_activity_;
+    uint32_t next_keepalive_time_;
+    uint32_t time_of_last_activity_;
     
     // Callback functions
     std::function<void(const std::string& sdp)> on_ice_candidate_;
